# resources/platform-protocol-specs/worker-protocol-api.openapi.yaml
openapi: 3.1.0
info:
  title: durable-workflow.v2.worker-protocol-api
  version: "30"
  summary: Durable Workflow worker-plane HTTP+JSON API
  description: >
    Normative OpenAPI specification for the worker-plane HTTP+JSON API:
    worker registration, graceful worker deregistration, worker heartbeat,
    worker-session lifecycle,
    long-poll task acquisition, task history paging, task heartbeat,
    task completion, and task failure. Long-poll timing, cancellation,
    and lease semantics are specified by durable-workflow.v2.worker-protocol-stream.
    The server advertises its current 1.N version and accepts request headers
    from the same major whose minor is less than or equal to N.
...
servers:
  - url: /api
security:
  - durableWorkerAuth: []
paths:
...
  /worker/workflow-tasks/{taskId}/local-activities/prepare:
    post:
      operationId: prepareLocalActivity
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Commit callback admission before application code starts. Retry an identical preparation with the same workflow claim and SDK attempt nonce.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalPrepareRequest" }
      responses:
        "200":
          description: The durable prepare result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalPreparationResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
...
  /worker/workflow-tasks/{taskId}/local-activities/recover:
    post:
      operationId: recoverLocalActivity
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        A replacement workflow claim recovers an expired original local attempt. Durable recovery records callback stop as unknown and preserves the original total timeout budget.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalRecoverRequest" }
      responses:
        "200":
          description: The durable recover result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalRecoveryResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
  /worker/workflow-tasks/{taskId}/local-activities/{attemptId}/control:
    post:
      operationId: controlLocalActivity
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Observe the original local attempt and optionally renew both workflow and attempt leases atomically. No application heartbeat is recorded. Accepted cancellation returns its original identity and deadline even after workflow claim takeover.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/PreparedLocalAttemptIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalControlRequest" }
      responses:
        "200":
          description: The durable control result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalControlResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
  /worker/workflow-tasks/{taskId}/local-activities/{attemptId}/heartbeat:
    post:
      operationId: heartbeatLocalActivity
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Record an application heartbeat and bounded progress under the original prepared claim. Update the heartbeat timeout without renewing either lease or moving fixed execution or root cleanup deadlines. Stale workers can observe cancellation but cannot record a heartbeat.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/PreparedLocalAttemptIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalHeartbeatRequest" }
      responses:
        "200":
          description: The durable heartbeat or stop result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalControlResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
  /worker/workflow-tasks/{taskId}/local-activities/{attemptId}/outcome:
    post:
      operationId: recordLocalActivityOutcome
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Commit an original callback outcome under its issued authority. An identical committed receipt remains readable after takeover. A stale worker cannot publish a fresh result.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/PreparedLocalAttemptIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalOutcomeRequest" }
      responses:
        "200":
          description: The durable outcome result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalOutcomeResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
  /worker/workflow-tasks/{taskId}/local-activities/{attemptId}/acknowledge-cancellation:
    post:
      operationId: acknowledgeLocalActivityCancellation
      tags: [workflow-tasks]
      x-durable-workflow-minimum-protocol-version: "1.20"
      x-durable-workflow-worker-capability: prepared_local_activities
      description: >-
        Report only after the original supervisor has stopped and joined the callback. A surviving supervisor may report after registration drain or claim takeover without gaining result or lease authority.
      parameters:
        - $ref: "#/components/parameters/WorkerProtocolVersionHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/PreparedLocalAttemptIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreparedLocalStopRequest" }
      responses:
        "200":
          description: The durable acknowledge-cancellation result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WorkerEnvelope"
                  - $ref: "#/components/schemas/PreparedLocalStopResult"
        "404": { $ref: "#/components/responses/WorkerError" }
        "409": { $ref: "#/components/responses/WorkerError" }
        "422": { $ref: "#/components/responses/WorkerError" }
        "429": { $ref: "#/components/responses/WorkerError" }
        "503": { $ref: "#/components/responses/WorkerServiceUnavailable" }
...
components:
  securitySchemes:
    durableWorkerAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    WorkerProtocolVersionHeader:
      name: X-Durable-Workflow-Protocol-Version
      in: header
      required: true
      description: >
        Worker-selected protocol version. The default server advertises 1.19
        and accepts same-major versions whose minor is no newer than 19.
        Missing, malformed, different-major, and ahead-of-server values fail closed.
      schema: { $ref: "#/components/schemas/AcceptedWorkerProtocolRequestVersion" }
    TaskIdPath:
      name: taskId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    PreparedLocalAttemptIdPath:
      name: attemptId
      in: path
      required: true
      description: Canonical durable attempt ID returned by preparation, distinct from worker_attempt_id.
      schema: { type: string, minLength: 1 }
...
  responses:
...
    WorkerError:
      description: Machine-readable worker-protocol error envelope.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkerError" }
    WorkerServiceUnavailable:
      description: >
        Temporary database connection loss returns backend_unavailable with an
        unknown operation outcome. Retry with bounded backoff and the identical
        worker, queue and poll_request_id; task=null does not prove that no task
        was claimed. Registration and worker heartbeat likewise do not acknowledge or
        reject an operation whose result is unknown. A workflow-task heartbeat
        carries task_id, lease_owner, and workflow_task_attempt so a worker can
        retry the same lease fence; a later 409 still revokes that lease. An
        activity completion carries task_id, activity_attempt_id and lease_owner
        to retry the same immutable result. Other
        503 errors retain their own contracts and are not automatically retryable. Optional
        storage admission returns storage_pressure or storage_admission_unavailable
        before admitting a request or a new poll claim. Retain the poll ID on
        retry: refusal of this attempt does not describe an earlier outcome.
      headers:
        X-Durable-Workflow-Protocol-Version:
          schema: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/WorkerBackendUnavailable"
              - $ref: "#/components/schemas/WorkerStorageAdmissionRefusal"
              - $ref: "#/components/schemas/WorkerError"
...
  schemas:
...
    WorkerEnvelope:
      type: object
      additionalProperties: true
      required: [protocol_version, server_capabilities]
      properties:
        protocol_version: { $ref: "#/components/schemas/AdvertisedWorkerProtocolVersion" }
        server_capabilities: { $ref: "#/components/schemas/WorkerServerCapabilities" }
...
    PreparedLocalRecoverRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalClaimRequest"
        - type: object
          required: [sequence, descriptor]
          properties:
            sequence: { type: integer, minimum: 1 }
            descriptor: { $ref: "#/components/schemas/PreparedLocalDescriptor" }
    PreparedLocalPrepareRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalRecoverRequest"
        - type: object
          required: [worker_attempt_id]
          properties:
            worker_attempt_id:
              type: string
              minLength: 1
              maxLength: 255
              description: Stable SDK nonce for this callback admission. It is distinct from the durable activity_attempt_id.
...
    PreparedLocalControlRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalClaimRequest"
        - type: object
          properties:
            renew_lease: { type: boolean, default: false }
    PreparedLocalHeartbeatRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalClaimRequest"
        - type: object
          properties:
            progress:
              type: [object, "null"]
              additionalProperties: false
              properties:
                message: { type: string, minLength: 1, maxLength: 280 }
                current: { type: [number, string], description: Nonnegative finite number or numeric string. }
                total: { type: [number, string], description: Nonnegative finite number or numeric string. }
                unit: { type: string, minLength: 1, maxLength: 64 }
                details: { type: object, maxProperties: 20 }
...
    PreparedLocalOutcomeRequest:
      allOf:
        - $ref: "#/components/schemas/PreparedLocalClaimRequest"
        - type: object
          required: [report]
          properties:
            report:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [outcome, result]
                  properties:
                    outcome: { const: completed }
                    result:
                      oneOf:
                        - { type: string, minLength: 1 }
                        - { type: object, required: [codec], properties: { codec: { const: avro } } }
                    payload_codec: { const: avro }
                - type: object
                  additionalProperties: false
                  required: [outcome, message]
                  properties:
                    outcome: { const: failed }
                    message: { type: string }
                    exception_type: { type: string, minLength: 1, maxLength: 255 }
                    non_retryable: { type: boolean, default: false }
                - type: object
                  additionalProperties: false
                  required: [outcome]
                  properties:
                    outcome: { const: timed_out }
...
    PreparedLocalPreparationResult:
      type: object
      required: [prepared, duplicate, reason, activity_execution_id, activity_attempt_id, worker_attempt_id, attempt_number, workflow_task_id, workflow_task_attempt, lease_owner, lease_expires_at, server_time]
      properties:
        prepared: { const: true }
        duplicate: { type: boolean }
        reason: { type: "null" }
        activity_execution_id: { type: string, minLength: 1 }
        activity_attempt_id: { type: string, minLength: 1 }
        worker_attempt_id: { type: string, minLength: 1 }
        attempt_number: { type: integer, minimum: 1 }
        workflow_task_id: { type: string, minLength: 1 }
        workflow_task_attempt: { type: integer, minimum: 1 }
        lease_owner: { type: string, minLength: 1 }
        lease_expires_at: { type: string, format: date-time }
        server_time: { type: string, format: date-time }
        start_to_close_deadline_at: { type: [string, "null"], format: date-time }
        schedule_to_close_deadline_at: { type: [string, "null"], format: date-time }
        heartbeat_deadline_at: { type: [string, "null"], format: date-time }
        cancellation_cleanup: { $ref: "#/components/schemas/PreparedLocalCleanupSnapshot" }
...
    PreparedLocalRecoveryResult:
      type: object
      required: [recovered, duplicate, reason, event_id, event_type, callback_stop_state, claim_released, created_task_ids]
      properties:
        recovered: { const: true }
        duplicate: { type: boolean }
        reason: { type: "null" }
        event_id: { type: string, minLength: 1 }
        event_type: { type: string, enum: [ActivityFailed, ActivityTimedOut, ActivityRetryScheduled] }
        callback_stop_state: { const: unknown }
        claim_released: { type: boolean }
        created_task_ids: { type: array, items: { type: string } }
    PreparedLocalControlResult:
      type: object
      required: [active, renewed, stop_required, reason, activity_execution_id, activity_attempt_id, workflow_task_id, workflow_task_attempt, server_time]
      properties:
        active: { type: boolean }
        renewed: { type: boolean }
        stop_required: { type: boolean }
        reason: { type: [string, "null"] }
        activity_execution_id: { type: string, minLength: 1 }
        activity_attempt_id: { type: string, minLength: 1 }
        workflow_task_id: { type: string, minLength: 1 }
        workflow_task_attempt: { type: integer, minimum: 1 }
        server_time: { type: string, format: date-time }
        heartbeat_recorded: { type: boolean }
        heartbeat_history_event_id: { type: [string, "null"] }
        cancellation_cleanup: { $ref: "#/components/schemas/PreparedLocalCleanupSnapshot" }
        cancellation_request:
          type: object
          required: [request_id, root_request_id, cleanup_deadline_at]
          properties:
            request_id: { type: string, minLength: 1 }
            root_request_id: { type: string, minLength: 1 }
            cleanup_deadline_at: { type: string, format: date-time }
...